refactor: create utils file in template
