+++ packages/untp-test-suite/src/templates/index.ts
@@ -1,2 +1,2 @@
-export * from './getTemplateName.js';
-export * from './generateFinalMessage.js';
+export * from './mapper.js';
+export * from './utils.js';
