+++ packages/untp-test-suite/src/templates/utils.ts
@@ -1,5 +1,12 @@
 import { ErrorObject } from 'ajv';
-import { IValidatedCredentials, TemplateName } from '../core/types/index.js';
+import {
+  IValidatedCredentials,
+  TemplateName,
+  FinalStatus,
+  ICredentialTestResult,
+  IFinalReport,
+  TestSuiteMessage,
+} from '../core/types/index.js';
 
 export const getTemplateName = (testSuiteResult: IValidatedCredentials): string => {
   if (testSuiteResult.errors === null) {
@@ -19,4 +26,24 @@ export const getTemplateName = (testSuiteResult: IValidatedCredentials): string 
   } else {
     return TemplateName.pass;
   }
+};
+
+export const generateFinalMessage = (credentials: ICredentialTestResult[]): IFinalReport => {
+  let finalMessage = TestSuiteMessage.Pass;
+  let finalStatus: FinalStatus = FinalStatus.pass;
+
+  if (credentials.some((credential) => credential.result === FinalStatus.warn)) {
+    finalMessage = TestSuiteMessage.Warning;
+    finalStatus = FinalStatus.warn;
+  }
+
+  if (credentials.some((credential) => credential.result === FinalStatus.fail)) {
+    finalMessage = TestSuiteMessage.Fail;
+    finalStatus = FinalStatus.fail;
+  }
+
+  return {
+    finalStatus,
+    finalMessage,
+  };
 };
